tracing: fix msdu_id processing pktlog_tx_msdu_id
Whenever trace record is started in middle of data flow, the
msdu_len_tbl is not fully populated. For the initial MSDU_ID log
types, msdu_id can not be found in msdu_len_tbl. If the entries
are not found set the msdu_len to 0. This will also fix below
errors.

  File "ath10k_pktlog.py", line 297, in <lambda>
    ath10k_htt_pktlog_handler(pevent, *args))
  File "ath10k_pktlog.py", line 238, in ath10k_htt_pktlog_handler
    pktlog_tx_msdu_id(buf)
  File "ath10k_pktlog.py", line 218, in pktlog_tx_msdu_id
    msdu_len = msdu_len_tbl[msdu_id]