tracing: extend plugin to support pktlog for 10.4 FW
Currently tracing plugin works only for QCA988X chipsets. This patch
extend the support to generate pktlog for QCA99X0.

With the ath10k commit "ath10k: add hw_rev to trace events to support
pktlog" we now have a new attribute passed with tracing events
"hw_type" that helps to determine the chipset used for taking the trace
points. This is helpful in getting FW variant used thereby generating
pktlog.dat accordingly.

To support backward compatibility for older versions of ath10k, assume the
default value of the hw_type, if not provided, to be HW_QCA988X (0).